# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple api server using the gin web framework.",
        "title": "Promo Code API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/api/v1",
    "paths": {
        "/add": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves promo code by name",
                "parameters": [
                    {
                        "description": "Example Data",
                        "name": "name,quantityAllocated,quantityAvailable,dateFrom,dateTo,amount",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Promocode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Promocode"
                        }
                    }
                }
            }
        },
        "/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves all promo codes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Promocode"
                        }
                    }
                }
            }
        },
        "/delete/{name}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Update promo code by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/update/{name}": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "summary": "Update promo code by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        }
    },
    "definitions": {
        "models.Promocode": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "format": "float64",
                    "example": 5.54
                },
                "dateFrom": {
                    "type": "string",
                    "format": "string",
                    "example": "1/12"
                },
                "dateTo": {
                    "type": "string",
                    "format": "string",
                    "example": "5/12"
                },
                "name": {
                    "type": "string",
                    "format": "string",
                    "example": "barone"
                },
                "quantityAllocated": {
                    "type": "integer",
                    "format": "int64",
                    "example": 12
                },
                "quantityAvailable": {
                    "type": "integer",
                    "format": "int64",
                    "example": 5
                }
            }
        }
    }
}